Makenek script breaks when parsing "_" in compiler names
Apparently, makenek script breaks when parsing "_" in compiler
names. Nek5000 does not need "_r" compilers as it does not
work with OpenMP threads. I think this is a breaking change
and will affect MFEM builds with threads. I wil work on to
fix the makenek script to handle "_" in compiler names.